[Tests] Remove Tests/ case duplicate from upmerge
